Add long overloads for methods

+++ micrometer-core/src/main/java/io/micrometer/core/instrument/MockClock.java
@@ -24,7 +24,7 @@ public class MockClock implements Clock {
 
     // has to be non-zero to prevent divide-by-zeroes and other weird math results based
     // on the clock
-    private long timeNanos = (long) TimeUtils.millisToUnit(1, TimeUnit.NANOSECONDS);
+    private long timeNanos = TimeUtils.millisToUnit(1, TimeUnit.NANOSECONDS);
 
     @Override
     public long monotonicTime() {
